requirements.txt
Authlib==1.2.1
blinker==1.6.2
click==8.1.7
flask-smorest==0.42.2
Flask-Migrate==4.0.5
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==1.10.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.18.0
gunicorn==20.1.0
numpy==1.26.0
pandas==2.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7
